+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -14,7 +14,7 @@ however, it has to be formatted properly to pass verification tests.
 - Fix UI sometimes ignoring the first mouse click event after losing and regaining focus ([case ISXB-127](https://issuetracker.unity3d.com/product/unity/issues/guid/ISXB-127).
 - Fixed issue when using MultiplayerEventSystems where the visual state of UI controls would change due to constant toggling of CanvasGroup.interactable on and off ([case ISXB-112](https://issuetracker.unity3d.com/product/unity/issues/guid/ISXB-112)).
 - Fixed an issue where the Input Action asset icon would not be visible during asset creation ([case ISXB-6](https://issuetracker.unity3d.com/product/unity/issues/guid/ISXB-6)).
-
+- Fixed an issue where setting an internal feature flag using IOCTL to communicate with native backend to `false` after being `true` would not have any effect (case ISXB-226). 
 
 ## [1.4.1] - 2022-05-30